requirements.txt
imea==0.3.5
matplotlib==3.4.3
numpy==2.2.1
Pillow==11.1.0
pycocotools==2.0.8
requests==2.31.0
torch==2.5.1
torchvision==0.20.1
transformers==4.47.1
